third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/iframe-javascript-url-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Ciframe%20src=javascript:alert(document.domain)%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
